[#30] Rename: userImage 컴포넌트 ui 폴더로 이동
